feat: add next token ID tracking
Have added a counter that keeps track of next available token ID for minting. This will help with the UI integration allowing us to fetch the next known free token ID for new minting.

This logic also sets the token ID new token mints to default to the next available token ID, meaning if ID 10 is attempted to be minted new, but only tokens with ID 1-4 have currently been minted, the ID of 5 will be used.